Mine clinic

+++ app/Http/Controllers/AppointmentController.php
@@ -90,11 +90,18 @@ class AppointmentController extends Controller
             session()->flash('success','Appointment has been declined');
             AppointmentDeclined::dispatch($appointment);
         } else if (intval($appointment->status) === Appointment::DONE_STATUS) {
-            if ($appointment->patient->has_medical_aid) {
-                SendInvoiceToMedicalAid::dispatch($appointment, $appointment->patient->medicalAid, $appointment->patient->medicalAid[0]->medical_email_address);
+            if ($appointment->appointmentAssessment->count() > 0) {
+                if ($appointment->patient->has_medical_aid) {
+                    SendInvoiceToMedicalAid::dispatch($appointment, $appointment->patient->medicalAid, $appointment->patient->medicalAid[0]->medical_email_address);
+                } else {
+                    SendInvoiceToPatient::dispatch($appointment);
+                }
             } else {
-                SendInvoiceToPatient::dispatch($appointment);
+                $appointment->status = Appointment::ACCEPTED_STATUS;
+                $appointment->save();
+                session()->flash('success','Please select all consultation(s) for this appointment');
             }
+
         }
 
         return redirect()->back();
@@ -108,10 +115,14 @@ class AppointmentController extends Controller
      */
     public function destroy(Appointment $appointment)
     {
-        SendDeletedAppointment::dispatch($appointment);
-        session()->flash('success','Appointment has been deleted');
+        if ($appointment->status !== Appointment::ACCEPTED_STATUS || $appointment->status !== Appointment::DONE_STATUS) {
+            SendDeletedAppointment::dispatch($appointment);
+            session()->flash('success','Appointment has been deleted');
+            $appointment->delete();
+        } else {
+            session()->flash('error','Appointment cannot be deleted');
+        }
 
-        $appointment->delete();
 
         return redirect()->back();
     }
